backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
amadeus==12.0.0
annotated-types==0.7.0
anthropic==0.50.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
azure-ai-documentintelligence==1.0.2
azure-ai-inference==1.0.0b9
azure-core==1.33.0
azure-identity==1.21.0
babel==2.17.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
cobble==0.1.4
coloredlogs==15.0.1
courlan==1.3.2
cryptography==44.0.2
dateparser==1.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-api-core==2.24.2
google-auth==2.40.2
google-geo-type==0.3.12
google-maps-places==0.2.1
google-maps-routing==0.6.15
googleapis-common-protos==1.70.0
googlemaps==4.10.0
gotrue==2.12.0
greenlet==3.2.0
grpcio==1.71.2
grpcio-status==1.71.2
h11==0.14.0
h2==4.2.0
hf-xet==1.1.1
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonref==1.1.0
jusText==3.0.2
loguru==0.7.3
lxml==5.3.2
lxml_html_clean==0.4.2
magika==0.6.1
Mako==1.3.10
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.0.2
MarkupSafe==3.0.2
mcp==1.9.1
mcp-amadeus==0.0.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
olefile==0.47
ollama==0.4.8
onnxruntime==1.21.0
openai==1.77.0
openpyxl==3.1.5
opentelemetry-api==1.35.0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pdfminer.six==20250416
pillow==11.2.1
playwright==1.51.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.7
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.5
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
sqlmodel==0.0.24
sse-starlette==2.3.5
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.13.3
textstat==0.7.7
threadpoolctl==3.6.0
tiktoken==0.9.0
tld==0.13
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.51.3
typer==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==14.2
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
yarl==1.20.0
youtube-transcript-api==1.0.3
zipp==3.21.0
